requirements.txt
# Core AI Framework
crewai>=1.9.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Web Framework
streamlit==1.31.0
flask==3.0.1

# ML Model Management
joblib==1.3.2

# Logging & Monitoring
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest>=8.0.0
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
# Core AI Framework
crewai==0.28.0
crewai[tools]==0.28.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Web Framework
streamlit==1.31.0
flask==3.0.1

# ML Model Management
joblib==1.3.2

# Logging & Monitoring
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
